first version with music search working, also it is playable

+++ NeoAlexandriaApp/app/src/main/java/com/example/neo_alexandria_app/DataModels/Resource.java
@@ -9,7 +9,7 @@ public class Resource {
     int commentCount;
     int saveCount;
     boolean isSaved;
-    double rating;
+    float rating;
 
     public Resource () {
 
@@ -47,7 +47,7 @@ public class Resource {
         return isSaved;
     }
 
-    public double getRating() {
+    public float getRating() {
         return rating;
     }
 }
